clean up, add company in own file
