+++ contract_isp_invoice/__init__.py
@@ -21,6 +21,7 @@
 ##############################################################################
 
 from . import (
+    company,
     contract,
     invoice,
     wizard,
